batch_processing/football/requirements.txt
requests
pandas
google-cloud-bigquery
google-cloud-aiplatform
google-cloud-storage
db-dtypes
scikit-learn
joblib
google-cloud-bigquery-storage